requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django>=5.0,<6.0
django-allauth==65.13.1
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
gunicorn==21.2.0
dj-database-url
